@@ -8,10 +8,11 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0",
+        "php": ">=5.4.0",
         "guzzlehttp/guzzle": "~4.0"
     },
     "require-dev": {
+        "mockery/mockery": "0.9.*",
         "phpunit/phpunit": "3.7.*"
     },
     "autoload": {
